added long read mode

+++ src/cloud_server.h
@@ -10,7 +10,7 @@
 #include "networking.h"
 #include "cloud_common.h"
 
-#define MAX_READ_BLOCK_SIZE (1024 * 1024 * 32) // 32 MiB
+#define MAX_READ_BLOCK_SIZE uint32_t(1024 * 1024 * 32) // 32 MiB
 
 class CloudConfig {
 public:
